+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Specific vocabulary versions
 - Several export formats for vocabularies: xlsx, json, ttl and and xml (linked data)
 - Keyword mapping no longer removes matched keywords from tag_string field
+- Change sorting of specific nodes in filter tree export used by frontend
 
 ## [1.2.0] - 2022-11-11